// src/pages/Marketing/MainPage/data.d.ts
export interface HotFundListIndex {
  fundCode: string;
  fundName: string;
  purchaseCumCnt: string;
  purchaseCumCntRadio: string;
  purchaseAmt: string;
  redeemAmt: string;
  netAmt: string;
  stockAmt: string;
  yearHeadInr: string;
}

export interface KeySalesListIndex {
  agencyCode: string;
  agencyName: string;
  purchaseAmt: string;
  redeemAmt: string;
  netAmt: string;
  stockAmt: string;
  holdNetAmt: string;
}